unified variadic arguments and results
