+++ elrond-wasm/src/storage/mappers/mod.rs
@@ -10,8 +10,8 @@ mod vec_mapper;
 pub use linked_list_mapper::LinkedListMapper;
 pub use map_mapper::MapMapper;
 pub use map_storage_mapper::MapStorageMapper;
-pub use mapper::StorageMapper;
 pub use mapper::StorageClearable;
+pub use mapper::StorageMapper;
 pub use set_mapper::SetMapper;
 pub use single_value_mapper::SingleValueMapper;
 pub use user_mapper::UserMapper;
