+++ elrond-wasm/src/types/io/optional_arg.rs
@@ -1,14 +1,26 @@
 use crate::abi::{TypeAbi, TypeDescriptionContainer};
-use crate::io::{ArgId, DynArg, DynArgInput};
+use crate::io::{ArgId, ContractCallArg, DynArg, DynArgInput};
+use crate::types::{ArgBuffer, SCError};
+use crate::{api::EndpointFinishApi, EndpointResult};
 use alloc::string::String;
 use elrond_codec::TopDecodeInput;
 
-/// A smart contract argument that can be provided or not.
+/// A smart contract  or result that can be missing.
 /// If arguments stop before this argument, None will be returned.
+/// As an endpoint result, the contract decides if it produces it or not.
+///
+/// As a principle, optional arguments or results should come last,
+/// otherwise there is ambiguity as to how to interpret what comes after.
+#[must_use]
 pub enum OptionalArg<T> {
 	Some(T),
 	None,
 }
+
+/// It is just an alias for `OptionalArg`.
+/// In general we use `OptionalArg` for arguments and `OptionalResult` for results,
+/// but it is the same implementation for both.
+pub type OptionalResult<T> = OptionalArg<T>;
 
 impl<T> From<Option<T>> for OptionalArg<T> {
 	fn from(v: Option<T>) -> Self {
@@ -45,9 +57,44 @@ where
 	}
 }
 
+impl<FA, T> EndpointResult<FA> for OptionalArg<T>
+where
+	FA: EndpointFinishApi + Clone + 'static,
+	T: EndpointResult<FA>,
+{
+	#[inline]
+	fn finish(&self, api: FA) {
+		if let OptionalResult::Some(t) = self {
+			t.finish(api);
+		}
+	}
+}
+
+impl<T> ContractCallArg for &OptionalArg<T>
+where
+	T: ContractCallArg,
+{
+	#[inline]
+	fn push_async_arg(&self, serializer: &mut ArgBuffer) -> Result<(), SCError> {
+		if let OptionalArg::Some(t) = self {
+			t.push_async_arg(serializer)?;
+		}
+		Ok(())
+	}
+}
+
+impl<T> ContractCallArg for OptionalArg<T>
+where
+	T: ContractCallArg,
+{
+	fn push_async_arg(&self, serializer: &mut ArgBuffer) -> Result<(), SCError> {
+		(&self).push_async_arg(serializer)
+	}
+}
+
 impl<T: TypeAbi> TypeAbi for OptionalArg<T> {
 	fn type_name() -> String {
-		let mut repr = String::from("OptionalArg<");
+		let mut repr = String::from("optional<");
 		repr.push_str(T::type_name().as_str());
 		repr.push('>');
 		repr
